requirements.txt
pytorch
transformers
sklearn
pandas
tensorboard
matplotlib
yaml
munch
wandb
sentencepiece